+++ facebook_auth_platform_interface/lib/src/access_token.dart
@@ -1,8 +1,58 @@
 const maxMillisecondsSinceEpoch = 8640000000000000;
 const minMillisecondsSinceEpoch = -8640000000000000;
 
+enum AccessTokenType { classic, limited }
+
+abstract class AccessToken {
+  final String tokenString;
+  final AccessTokenType type;
+
+  AccessToken({
+    required this.tokenString,
+    required this.type,
+  });
+
+  Map<String, dynamic> toJson();
+}
+
+class LimitedToken extends AccessToken {
+  final String userId;
+  final String userName;
+  final String userEmail;
+  final String nonce;
+
+  LimitedToken({
+    required this.userId,
+    required this.userName,
+    required this.userEmail,
+    required this.nonce,
+    super.type = AccessTokenType.limited,
+    required super.tokenString,
+  });
+
+  factory LimitedToken.fromJson(Map<String, dynamic> json) {
+    return LimitedToken(
+      userId: json['userId'],
+      userName: json['userName'],
+      userEmail: json['userEmail'],
+      tokenString: json['token'],
+      nonce: json['nonce'],
+    );
+  }
+
+  @override
+  Map<String, dynamic> toJson() => {
+        'type': type.name,
+        'userId': userId,
+        'tokenString': tokenString,
+        'nonce': nonce,
+        'userEmail': userEmail,
+        'userName': userName,
+      };
+}
+
 /// Class that contains the facebook access token data
-class AccessToken {
+class ClassicToken extends AccessToken {
   /// DateTime with the expires date of this token
   final DateTime expires;
 
@@ -11,9 +61,6 @@ class AccessToken {
 
   /// the facebook user id
   final String userId;
-
-  /// token provided by facebook to make api calls to the GRAPH API
-  final String token;
 
   /// the facebook application Id
   final String applicationId;
@@ -33,23 +80,24 @@ class AccessToken {
   /// DateTime with the date at which user data access expires
   final DateTime dataAccessExpirationTime;
 
-  AccessToken({
+  ClassicToken({
     required this.declinedPermissions,
     required this.grantedPermissions,
     required this.userId,
     required this.expires,
     required this.lastRefresh,
-    required this.token,
+    required super.tokenString,
     required this.applicationId,
     this.graphDomain,
     required this.isExpired,
     required this.dataAccessExpirationTime,
+    super.type = AccessTokenType.classic,
   });
 
   /// convert the data provided for the platform channel to one instance of AccessToken
   ///
   /// [json] data returned by the platform channel
-  factory AccessToken.fromJson(Map<String, dynamic> json) {
+  factory ClassicToken.fromJson(Map<String, dynamic> json) {
     late final DateTime dataAccessExpirationTime;
 
     if (json['dataAccessExpirationTime'] is String) {
@@ -62,9 +110,9 @@ class AccessToken {
       );
     }
 
-    return AccessToken(
+    return ClassicToken(
       userId: json['userId'],
-      token: json['token'],
+      tokenString: json['token'],
       expires: DateTime.fromMillisecondsSinceEpoch(
         json['expires'].clamp(
           minMillisecondsSinceEpoch,
@@ -86,9 +134,11 @@ class AccessToken {
   }
 
   /// convert this instance to one Map
+  @override
   Map<String, dynamic> toJson() => {
+        'type': type.name,
         'userId': userId,
-        'token': token,
+        'tokenString': tokenString,
         'expires': expires.millisecondsSinceEpoch,
         'lastRefresh': lastRefresh.millisecondsSinceEpoch,
         'applicationId': applicationId,